+++ src/components/ui/layout-components/table-filter-bar/table-filter-bar.tsx
@@ -2,6 +2,7 @@ import { ChangeEvent } from 'react'
 
 import { TrashOutline } from '@/assets/icons'
 import { Button, Slider, TabSwitcher, TextField } from '@/components/ui/primitives'
+import { useSearchParamUpdater } from '@/shared/hooks'
 import { Tab } from '@/shared/types/common'
 import { FlexContainer } from '@/shared/ui/flex-container'
 
@@ -10,7 +11,6 @@ import { cn } from './table-filter-bar.styles'
 type TableFilterBarProps = {
   max: number
   min: number
-  onClearFilters: () => void
   onSearchChange: (e: ChangeEvent<HTMLInputElement>) => void
   onSliderChange: (sliderRange: number[]) => void
   onTabChange: (tab: string) => void
@@ -19,17 +19,31 @@ type TableFilterBarProps = {
 export const TableFilterBar = ({
   max,
   min,
-  onClearFilters,
   onSearchChange,
   onSliderChange,
   onTabChange,
   search,
 }: TableFilterBarProps) => {
+  const updateSearchParam = useSearchParamUpdater()
+
   const tabs: Tab[] = [
     { title: 'All Decks', value: 'allDecks' },
     { title: 'My Decks', value: 'myDecks' },
     { title: 'Favorites', value: 'favorites' },
   ]
+
+  const clearFiltersHandler = () => {
+    updateSearchParam({
+      authorId: '',
+      currentPage: 1,
+      itemsPerPage: 10,
+      max: 100,
+      min: 0,
+      orderBy: '',
+      search: '',
+      tab: 'allDecks',
+    })
+  }
 
   return (
     <FlexContainer ai={'flex-end'} fd={'row'} gap={'24px'}>
@@ -46,7 +60,7 @@ export const TableFilterBar = ({
         tabs={tabs}
       />
       <Slider label={'Number of cards'} max={max} min={min} onRangeChange={onSliderChange} />
-      <Button className={cn.button} onClick={onClearFilters} variant={'secondary'}>
+      <Button className={cn.button} onClick={clearFiltersHandler} variant={'secondary'}>
         <TrashOutline />
         Clear Filter
       </Button>
